Implement changes to quorum data structure to include the addition of transaction body digests and datatype changes to resource testing digest.

+++ validator/quorum/validator.go
@@ -44,7 +44,7 @@ type vote struct {
 	Day                           uint8
 	Month                         uint8
 	Year                          uint8
-	PreviousResourceTestingDigest uint64
+	PreviousResourceTestingDigest uint32
 	PreviousSpectrumDigest        [32]byte
 	PreviousUniverseDigest        [32]byte
 	PreviousComputerDigest        [32]byte
